lane change logic
switch lane if there is a switch and the optimal lane of the next bend
after that is different from the one we're riding at.

this would need to check all the following bends after the next switch
to see which side would be optimal, instead of the piece that
immediately follows the switch.